requirements.txt
alembic==1.4.3
amqp==5.0.2
attrs==20.2.0
autopep8==1.5.4
bcrypt==3.2.0
billiard==3.6.3.0
blinker==1.4
boto3==1.16.25
botocore==1.19.25
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.1
Faker==6.4.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
infinity==1.5
iniconfig==1.1.1
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
mailchimp-marketing==3.0.25
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
qrcode==6.1
requests==2.25.0
rq==1.6.1
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
validators==0.18.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Components==0.10.4
WTFormValidation==1.0.0
zipp==3.4.0
